fix: change type of 'now' in process_packets to double for improved precision

+++ dpdk-23.11/app/raft-netelect/networking.c
@@ -213,7 +213,7 @@ void process_packets(void)
                 (struct ps_broadcast_packet *)payload;
 
             uint32_t peer = ps->node_id;
-            uint64_t now  = rte_get_tsc_cycles();
+            double now  = rte_get_tsc_cycles();
             
 
             // simple RTT measurement
